Fixed bugs with explicit abort (XABORT)
The explicit abort (XABORT) was not working properly. First, on issue
the ReorderBufferEntry::issue() was continued after flushing the pipeline
on XABORT, which caused erroneous behaviour. Second, in
TsxCache::handle_upper_interconnect() the check_tsx_invalidated()
function is called on XABORT, but it does not make sense, since it mixes
the TsxStatus and TsxMESICacheLineState data types up.