Fixed aliasing bug in TSX memory buffer
All transactional data accesses that map the same line in the
associative array structure are not aligned to their block offset,
i.e. they overwrite each others entries on write and return the
same data on read access. Applied an early fix for this issue
by setting the line size of the buffer to 1.